fix: move queries in query folder
